+++ diatom-cli/Cargo.toml
@@ -16,7 +16,3 @@ crossterm = { version = "0.24" }
 clap = { version = "4", features = ["derive"] }
 regex.workspace = true
 lazy_static.workspace = true
-
-[[bin]]
-name = "diatom"
-path = "src/main.rs"
